requirements.txt
django==3.2.16
djangorestframework==3.14.0
psycopg2==2.9.5
django-environ==0.9.0
django-cors-headers==3.13.0
Pillow==9.3.0
django-storages==1.13.1
django-ckeditor==6.3.2
django-coreapi == 2.3.0

# del video de autenticacion
djangorestframework-simplejwt==5.3.0
#djoser==2.2.0
PyJWT==2.6.0
pytz==2022.5
asgiref==3.5.2
pytz==2022.5

